Use correct name of operating system.
As of version 10.8 Mountain Lion, the name of the operating system was renamed from "Mac OS X" to just "OS X".

Mac* (MacBook, Mac Pro, etc.) is the name of the computer hardware. OS X is the name of its operating system.

Similarly, i* (iPhone, iPad, etc.) is the name of the mobile device hardware. iOS is the name of its operating system.

I only changes instances of "Mac OS" and "Mac OS X" to "OS X", since that was referring to software (by an incorrect name). I kept instances of just "Mac" (e.g. "Mac app") because that refers to the hardware (by a correct name), and it's not up to me to change that but up to the app developer.

Sources:
- https://en.wikipedia.org/wiki/OS_X_Mountain_Lion#cite_note-PR-16-02-2
- https://en.wikipedia.org/wiki/OS_X_Mountain_Lion#cite_note-dropsmac-4